Move ContentType sub-namespace classes
This subset of the current classes should sit within the Models\ContentType sub-namespace. This also removes the redundant sub-namespace of the class name.
Classes depending on these ones have been updated.
