+++ src/lib/models/content_type.php
@@ -4,11 +4,8 @@
  * */
 namespace Contentstack\Stack\ContentType;
 
-use Contentstack\Stack\ContentType\Entry\Entry;
-use Contentstack\Stack\ContentType\Query\Query;
-
-require_once __DIR__.'/entry.php';
-require_once __DIR__.'/query.php';
+use Contentstack\Models\ContentType\Entry;
+use Contentstack\Models\ContentType\Query;
 
 /*
  * Class ContentType
